Add logic to allow deleting of cart items

+++ models/product.js
@@ -51,7 +51,10 @@ module.exports = class Product {
     static deleteById(id) {
         getProductsFromFile(products => {
             const product = products.find(prod => prod.id === id);
-            const updatedProducts = products.filter(prod => prod.id !== id);
+			const updatedProducts = products.filter(prod => prod.id !== id);
+			if (!product) {
+				return;
+			}
             fs.writeFile(p, JSON.stringify(updatedProducts), err => {
                 if (!err) {
                     Cart.deleteProduct(id, product.price);
